Eliminando erros na funcao pesquisar

+++ ControleAtleta/src/controleatleta/Atleta.java
@@ -8,7 +8,7 @@ public class Atleta {
     ArrayList <String> telefones;
     private String nome;
     private Date dataNascimento;
-    private Double idade;
+    private int idade;
     private Endereco endereco;
     private Double altura;
     private Double peso;
@@ -47,11 +47,11 @@ public class Atleta {
         this.dataNascimento = dataNascimento;
     }
 
-    public Double getIdade() {
+    public int getIdade() {
         return idade;
     }
 
-    public void setIdade(Double idade) {
+    public void setIdade(int idade) {
         this.idade = idade;
     }
